Rename file to PrizeBatchList
