+++ src/views/Prizes/index.ts
@@ -15,4 +15,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from "./PrizeBatches";
+export * from "./PrizeBatchList";
